All logger patching in one file rawk_log/patch_logger
