+++ lib/rawk_log.rb
@@ -3,13 +3,6 @@ require "rawk_log/version"
 # Don't load automatically
 #require "rawk_log/command"
 
-if defined?(Logger)
-  require "rawk_log/patch_logger"
-end
-
-if defined?(ActiveSupport::BufferedLogger)
-  require "rawk_log/patch_activesupport_bufferedlogger"
-end
 
 module RawkLog
   # Everything loaded in requires above
